[feat] 그룹채팅방 actionInfo 모달창 컴포넌트 분리, page.detail 타입에러 해결
